+++ lizard_map/views.py
@@ -230,15 +230,8 @@ class MapMixin(object):
         return False
 
     def background_maps(self):
-        if self.backgrounds:
-            return self.backgrounds
-        logger.warn("No background maps are active. Taking default.")
-        return [BackgroundMap(
-                name='Default map',
-                default=True,
-                active=True,
-                layer_type=BackgroundMap.LAYER_TYPE_OSM,
-                layer_url=DEFAULT_OSM_LAYER_URL), ]
+        """Return current background maps, or the default if there are none."""
+        return self.backgrounds or BackgroundMap.default_maps()
 
 
 class CollageMixin(object):